refactor(session): 拆出 trace 事件兼容读取 mixin 守住千行上限
trace_event_chunks.py 在取消释放修复后达 1021 行，超出后端 1000 行红线
（CI Large File Check 与前端 largeBackendModulesSource 结构测试双重强制）。
按 trace_chunk_rollback 的既有先例，把 _has_event_chunks /
read_trace_events_compat / read_trace_events_batch_compat 迁移到新模块
trace_event_read.py，TraceStorage 直接混入，行为不变；并补齐 ruff format。
